feature: add viewpost page and refactor comments
